Base,BaseGL: move StipplingProperty to BaseGL
