+++ modules/basegl/CMakeLists.txt
@@ -12,6 +12,8 @@ set(HEADER_FILES
     include/modules/basegl/datastructures/linesettings.h
     include/modules/basegl/datastructures/linesettingsinterface.h
     include/modules/basegl/datastructures/meshshadercache.h
+    include/modules/basegl/datastructures/stipplingsettings.h
+    include/modules/basegl/datastructures/stipplingsettingsinterface.h
     include/modules/basegl/datavisualizer/imagebackgroundvisualizer.h
     include/modules/basegl/datavisualizer/imagevisualizer.h
     include/modules/basegl/datavisualizer/meshvisualizer.h
@@ -76,6 +78,7 @@ set(HEADER_FILES
     include/modules/basegl/processors/volumeraycaster.h
     include/modules/basegl/processors/volumeslicegl.h
     include/modules/basegl/properties/linesettingsproperty.h
+    include/modules/basegl/properties/stipplingproperty.h
     include/modules/basegl/rendering/linerenderer.h
     include/modules/basegl/viewmanager.h
 )
@@ -90,6 +93,8 @@ set(SOURCE_FILES
     src/datastructures/linesettings.cpp
     src/datastructures/linesettingsinterface.cpp
     src/datastructures/meshshadercache.cpp
+    src/datastructures/stipplingsettings.cpp
+    src/datastructures/stipplingsettingsinterface.cpp
     src/datavisualizer/imagebackgroundvisualizer.cpp
     src/datavisualizer/imagevisualizer.cpp
     src/datavisualizer/meshvisualizer.cpp
@@ -154,6 +159,7 @@ set(SOURCE_FILES
     src/processors/volumeraycaster.cpp
     src/processors/volumeslicegl.cpp
     src/properties/linesettingsproperty.cpp
+    src/properties/stipplingproperty.cpp
     src/rendering/linerenderer.cpp
     src/viewmanager.cpp
 )
