+++ modules/basegl/include/modules/basegl/properties/linesettingsproperty.h
@@ -29,7 +29,7 @@
 #pragma once
 
 #include <modules/basegl/datastructures/linesettingsinterface.h>
-#include <modules/base/properties/stipplingproperty.h>
+#include <modules/basegl/properties/stipplingproperty.h>
 #include <inviwo/core/properties/compositeproperty.h>
 #include <inviwo/core/properties/boolproperty.h>
 #include <inviwo/core/properties/ordinalproperty.h>
