+++ modules/basegl/src/baseglmodule.cpp
@@ -86,6 +86,7 @@
 #include <modules/basegl/processors/volumeraycaster.h>
 #include <modules/basegl/processors/volumeslicegl.h>
 #include <modules/basegl/processors/volumeprocessing/volumeshader.h>
+#include <modules/basegl/properties/stipplingproperty.h>
 #include <modules/basegl/datavisualizer/volumeraycastvisualizer.h>
 #include <modules/basegl/datavisualizer/volumeslicevisualizer.h>
 #include <modules/basegl/datavisualizer/imagevisualizer.h>
@@ -102,6 +103,8 @@ namespace inviwo {
 BaseGLModule::BaseGLModule(InviwoApplication* app) : InviwoModule(app, "BaseGL") {
 
     basegl::addShaderResources(ShaderManager::getPtr(), {getPath(ModulePath::GLSL)});
+
+    registerProperty<StipplingProperty>();
 
     registerProcessor<AxisAlignedCutPlane>();
     registerProcessor<Background>();
