+++ modules/opengl/depends.cmake
@@ -1,6 +1,5 @@
 #--------------------------------------------------------------------
 # Dependencies for OpenGL module
 set(dependencies 
-	InviwoBaseModule
 )
 set(EnableByDefault ON)
